StandaloneMmPkg: CodeQL Fixes.
Makes changes to comply with alerts raised by CodeQL.

Most of the issues here fall into the following two categories:

1. Potential use of uninitialized pointer
2. Inconsistent integer width used in loop comparison

+++ StandaloneMmPkg/Library/FvLib/FvLib.c
@@ -167,7 +167,8 @@ FfsFindNextFile (
 
   FileOffset = (UINT32)((UINT8 *)FfsFileHeader - (UINT8 *)FwVolHeader);
 
-  while (FileOffset < (FvLength - sizeof (EFI_FFS_FILE_HEADER))) {
+  while ((UINT64)FileOffset < (FvLength - sizeof (EFI_FFS_FILE_HEADER))) {
+    // MU_CHANGE - CodeQL Change - comparison-with-wider-type
     //
     // Get FileState which is the highest bit of the State
     //
